+++ src/Agent.Worker/TaskRunner.cs
@@ -214,9 +214,15 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
                 // for security reasons execution of tasks' in this case should be skipped.
                 // Target task inputs could be injected into the decorator's tasks if the decorator has post-task-tasks or pre-task-tasks targets,
                 // such tasks will have names that start with __system_pretargettask_ or __system_posttargettask_.
-                if (this.IsInjectedTaskForTarget(Task.Name))
+                var taskDecoratorManager = HostContext.GetService<ITaskDecoratorManager>();
+                if (taskDecoratorManager.IsInjectedTaskForTarget(Task.Name) &&
+                    taskDecoratorManager.IsInjectedInputsContainsSecrets(inputs, out var inputsWithSecrets))
                 {
-                    this.CheckInjectedTasksInputs(inputs);
+                    var inputsForReport = taskDecoratorManager.GenerateTaskResultMessage(inputsWithSecrets);
+                    
+                    ExecutionContext.Result = TaskResult.Skipped;
+                    ExecutionContext.ResultCode = StringUtil.Loc("SecretsAreNotAllowedInInjectedTaskInputs", inputsForReport);
+                    return;
                 }
 
                 VarUtil.ExpandEnvironmentVariables(HostContext, target: inputs);
@@ -535,74 +541,5 @@ namespace Microsoft.VisualStudio.Services.Agent.Worker
             ExecutionContext.Output($"Help         : {taskDefinition.Data.HelpUrl ?? taskDefinition.Data.HelpMarkDown}");
             ExecutionContext.Output("==============================================================================");
         }
-
-        /// <summary>
-        /// Used to check if provided input value contain any secret
-        /// </summary>
-        /// <param name="inputValue">Value of input to check</param>
-        /// <returns>Returns `true` if provided string contain secret, otherwise `false`</returns>
-        private bool ContainsSecret(string inputValue)
-        {
-            string maskedString = HostContext.SecretMasker.MaskSecrets(inputValue);
-            return maskedString != inputValue;
-        }
-
-        /// <summary>
-        /// Used to get list of inputs in injected task that started with target_ prefix and contain secrets,
-        /// such inputs are autoinjected from target tasks
-        /// </summary>
-        /// <param name="inputs">Inputs presented as a dictionary with input name as key and input's value as the value of the corresponding key</param>
-        /// <returns>Returns list of inputs' names that contain secret values</returns>
-        private List<string> GetInputsWithSecrets(Dictionary<string, string> inputs)
-        {
-            var inputsWithSecrets = new List<string>();
-            foreach (var input in inputs)
-            { 
-                if (input.Key.StartsWith("target_") && this.ContainsSecret(input.Value))
-                {
-                    inputsWithSecrets.Add(input.Key);
-                }
-            }
-
-            return inputsWithSecrets;
-        }
-
-        /// <summary>
-        /// Used to check list of inputs in injected task that started with target_ prefix and contain secrets,
-        /// such inputs are autoinjected from target tasks
-        /// </summary>
-        /// <param name="inputs">Inputs presented as a dictionary with input name as key and input's value as the value of the corresponding key</param>
-        /// <returns>Returns list of inputs' names that contain secret values</returns>
-        private void CheckInjectedTasksInputs(Dictionary<string, string> inputs)
-        {
-            var inputsWithSecrets = this.GetInputsWithSecrets(inputs);
-
-            if (inputsWithSecrets.Count > 0)
-            {
-                string inputsForReport = string.Join(Environment.NewLine,
-                    inputsWithSecrets.Select(input => string.Join("\n", input)));
-
-                ExecutionContext.Result = TaskResult.Skipped;
-                ExecutionContext.ResultCode = StringUtil.Loc("SecretsAreNotAllowedInInjectedTaskInputs", inputsForReport);
-                return;
-            }
-        }
-
-        /// <summary>
-        /// Checks if current task is injected by decorator with posttargettask or pretargettask target
-        /// </summary>
-        /// <param name="taskName">Name of the task to check</param>
-        /// <returns>Returns `true` if task is injected by decorator for target task, otherwise `false`</returns>
-        private bool IsInjectedTaskForTarget(string taskName)
-        {
-            return taskName.StartsWith(InjectedTasksNamesPrefixes.PostTargetTask)
-                    || taskName.StartsWith(InjectedTasksNamesPrefixes.PreTargetTask);
-        }
-    }
-
-    internal static class InjectedTasksNamesPrefixes
-    {
-        public static readonly String PostTargetTask = "__system_posttargettask_";
-        public static readonly String PreTargetTask = "__system_pretargettask_";
     }
 }
